03. Integrando classes e definindo relacionamentos

+++ Musicas.cs
@@ -2,7 +2,7 @@ class Musica
 {
     public string Nome { get; set; }
     public string Artista { get; set; }
-    public int Duracao { get; set; }
+    public double Duracao { get; set; }
     public bool Disponivel { get; set; }
     public string DescricaoResumida => $"A música {Nome} pertence à banda {Artista}";
